Re-organized and cleaned up project
