+++ datasets/analysis-for-city/src/analysis_for_city/generate_damage_report.py
